+++ src/org/bds/executioner/QueueThreadAwsSqs.java
@@ -27,6 +27,7 @@ public class QueueThreadAwsSqs extends QueueThread {
 	public static String QUEUE_NAME_SUFFIX_DEBUG = "test_123456789";
 	public static boolean USE_QUEUE_NAME_DEBUG = true; // Fix queue name when running tests or debugging
 
+	// FIXME: IMPLEMENT COMMAND IN BDS_EXEC
 	private static final String OS_DELETE_QUEUE_COMMAND = "@aws_sqs_delete_queue"; // Command line to delete a queue
 	public static final int AWS_SQS_WAIT_TIME_SECONDS = 10; // Long polling parameter
